refactor(gameday): moved `gameday` to pkgs
